+++ CHANGES.md
@@ -19,7 +19,7 @@ SDK API changes ⚠️:
  -
 
 Build 🧱:
- -
+ - Remove dependency to org.greenrobot.eventbus library
 
 Other changes:
  -